+++ src/Geometry/Bound.hs
@@ -18,13 +18,13 @@ data Bound = Bound
 instance Semigroup Bound where
   (<>) (Bound (mxX1, mxY1) (mnX1, mnY1)) (Bound (mxX2, mxY2) (mnX2, mnY2)) = Bound (max mxX1 mxX2, max mxY1 mxY2) (min mnX1 mnX2, min mnY1 mnY2)
 
-maxX :: Bound -> Float
+maxX :: Bound -> Double
 maxX = fst . maxB
-maxY :: Bound -> Float
+maxY :: Bound -> Double
 maxY = snd . maxB
-minX :: Bound -> Float
+minX :: Bound -> Double
 minX = fst . minB
-minY :: Bound -> Float
+minY :: Bound -> Double
 minY = snd . minB
 
 overlap :: Bound -> Bound -> Maybe Bound
@@ -41,7 +41,7 @@ bound [v] = boundOne v
 bound ((x,y):vs) = Bound (max mxX x, max mxY y) (min mnX x, min mnY y)
   where (Bound (mxX, mxY) (mnX, mnY)) = bound vs
 
-bufferedBound :: [Position] -> Float -> Bound
+bufferedBound :: [Position] -> Double -> Bound
 bufferedBound vs b = Bound (maxB bnd |+ buff) (minB bnd |- buff)
   where
     bnd = bound vs
