+++ src/View.hs
@@ -10,7 +10,7 @@ import Struct
 data View c = View
   { structOrModel :: Either (Struct c) (Model c)
   , center :: Position
-  , zoom :: Float
+  , zoom :: Double
   }
 
 instance HasPos (View c) where
