+++ src/Wall.hs
@@ -9,7 +9,7 @@ import Point
 import Geometry.Vector
 import Geometry.Circle
 
-data Wall = VLine Float | HLine Float | Rock Circle deriving (Eq, Show)
+data Wall = VLine Double | HLine Double | Rock Circle deriving (Eq, Show)
 
 rock :: Position -> Radius -> Wall
 rock p = Rock . Circle p
@@ -18,7 +18,7 @@ wSide :: Wall -> Position -> Side
 wSide (VLine x) p = wSide' x $ fst p
 wSide (HLine y) p = wSide' y $ snd p
 wSide (Rock (Circle c r)) p = if distSq p c > r^2 then Out else In
-wSide' :: Float -> Float -> Side
+wSide' :: Double -> Double -> Side
 wSide' pl p = case compare pl p of
   LT -> In
   _ -> Out
